#431 :truck: rename mapper + move back to .service package
