+++ wls-broadcast-service/src/main/java/de/muenchen/oss/wahllokalsystem/broadcastservice/service/BroadcastService.java
@@ -2,7 +2,6 @@ package de.muenchen.oss.wahllokalsystem.broadcastservice.service;
 
 import de.muenchen.oss.wahllokalsystem.broadcastservice.domain.Message;
 import de.muenchen.oss.wahllokalsystem.broadcastservice.domain.MessageRepository;
-import de.muenchen.oss.wahllokalsystem.broadcastservice.rest.BroadcastDTOMapper;
 import de.muenchen.oss.wahllokalsystem.broadcastservice.rest.BroadcastMessageDTO;
 import de.muenchen.oss.wahllokalsystem.broadcastservice.rest.MessageDTO;
 import de.muenchen.oss.wahllokalsystem.broadcastservice.util.BroadcastExceptionKonstanten;
@@ -24,7 +23,7 @@ public class BroadcastService {
 
     private final MessageRepository messageRepo;
 
-    private final BroadcastDTOMapper broadcastMapper;
+    private final BroadcastMapper broadcastMapper;
 
     private final ExceptionFactory exceptionFactory;
