+++ wls-broadcast-service/src/test/java/de/muenchen/oss/wahllokalsystem/broadcastservice/rest/BroadcastDTOMapperTest.java
@@ -1,5 +1,6 @@
 package de.muenchen.oss.wahllokalsystem.broadcastservice.rest;
 
+import de.muenchen.oss.wahllokalsystem.broadcastservice.service.BroadcastMapper;
 import de.muenchen.oss.wahllokalsystem.broadcastservice.utils.TestdataFactory;
 import java.time.LocalDateTime;
 import java.util.Arrays;
@@ -11,7 +12,7 @@ import org.mapstruct.factory.Mappers;
 
 public class BroadcastDTOMapperTest {
 
-    private final BroadcastDTOMapper unitUnderTest = Mappers.getMapper(BroadcastDTOMapper.class);
+    private final BroadcastMapper unitUnderTest = Mappers.getMapper(BroadcastMapper.class);
 
     @Nested
     class ToDto {
