+++ wls-broadcast-service/src/test/java/de/muenchen/oss/wahllokalsystem/broadcastservice/service/BroadcastServiceTest.java
@@ -2,7 +2,6 @@ package de.muenchen.oss.wahllokalsystem.broadcastservice.service;
 
 import de.muenchen.oss.wahllokalsystem.broadcastservice.domain.Message;
 import de.muenchen.oss.wahllokalsystem.broadcastservice.domain.MessageRepository;
-import de.muenchen.oss.wahllokalsystem.broadcastservice.rest.BroadcastDTOMapper;
 import de.muenchen.oss.wahllokalsystem.broadcastservice.rest.BroadcastMessageDTO;
 import de.muenchen.oss.wahllokalsystem.broadcastservice.rest.MessageDTO;
 import de.muenchen.oss.wahllokalsystem.broadcastservice.util.BroadcastExceptionKonstanten;
@@ -35,7 +34,7 @@ class BroadcastServiceTest {
     private MessageRepository messageRepo;
 
     @Mock
-    private BroadcastDTOMapper broadcastMapper;
+    private BroadcastMapper broadcastMapper;
 
     @Mock
     ExceptionFactory exceptionFactory;
